Test fix for reshape and sqrt

+++ ivy_tests/test_ivy/test_frontends/test_torch/test_pointwise_ops.py
@@ -934,9 +934,7 @@ def test_torch_trunc(
 # sqrt
 @handle_cmd_line_args
 @given(
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric")
-    ),
+    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("float")),
     num_positional_args=helpers.num_positional_args(
         fn_name="ivy.functional.frontends.torch.sqrt"
     ),
